move encoder to within their serialization packages
