+++ exporter/opentelemetry-exporter-zipkin-json/tests/encoder/test_v2_json.py
@@ -15,7 +15,7 @@ import json
 
 from opentelemetry import trace as trace_api
 from opentelemetry.exporter.zipkin.encoder import NAME_KEY, VERSION_KEY
-from opentelemetry.exporter.zipkin.encoder.v2.json import JsonV2Encoder
+from opentelemetry.exporter.zipkin.json.v2 import JsonV2Encoder
 from opentelemetry.exporter.zipkin.node_endpoint import NodeEndpoint
 from opentelemetry.sdk import trace
 from opentelemetry.trace import SpanKind, TraceFlags
