+++ exporter/opentelemetry-exporter-zipkin-json/tests/test_zipkin_exporter.py
@@ -19,7 +19,7 @@ from unittest.mock import patch
 
 from opentelemetry import trace
 from opentelemetry.exporter.zipkin.encoder import Protocol
-from opentelemetry.exporter.zipkin.encoder.v2.json import JsonV2Encoder
+from opentelemetry.exporter.zipkin.json.v2 import JsonV2Encoder
 from opentelemetry.exporter.zipkin.json import DEFAULT_ENDPOINT, ZipkinExporter
 from opentelemetry.exporter.zipkin.node_endpoint import NodeEndpoint
 from opentelemetry.sdk.environment_variables import (
