+++ exporter/opentelemetry-exporter-zipkin-proto/tests/encoder/test_v2_protobuf.py
@@ -15,8 +15,8 @@ import ipaddress
 import json
 
 from opentelemetry.exporter.zipkin.encoder import NAME_KEY, VERSION_KEY
-from opentelemetry.exporter.zipkin.encoder.v2.protobuf import ProtobufEncoder
-from opentelemetry.exporter.zipkin.encoder.v2.protobuf.gen import zipkin_pb2
+from opentelemetry.exporter.zipkin.proto.v2 import ProtobufEncoder
+from opentelemetry.exporter.zipkin.proto.v2.gen import zipkin_pb2
 from opentelemetry.exporter.zipkin.node_endpoint import NodeEndpoint
 from opentelemetry.trace import SpanKind
