+++ exporter/opentelemetry-exporter-zipkin-proto/tests/test_zipkin_exporter.py
@@ -18,7 +18,7 @@ import unittest
 from unittest.mock import patch
 
 from opentelemetry import trace
-from opentelemetry.exporter.zipkin.encoder.v2.protobuf import ProtobufEncoder
+from opentelemetry.exporter.zipkin.proto.v2 import ProtobufEncoder
 from opentelemetry.exporter.zipkin.node_endpoint import NodeEndpoint
 from opentelemetry.exporter.zipkin.proto import (
     DEFAULT_ENDPOINT,
